src/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
httpx
gunicorn
reportlab
azure-storage-blob
azure-cosmos
pydantic[email]
openai
PyPDF2
pytz
azure-keyvault-secrets